NIFI-12506 Addrassing review comments
